+++ api/src/db/migrations/2021-11-05_multiple-collaborations.js
@@ -1,0 +1,7 @@
+const sequelize = require("../sequelize");
+
+sequelize.query(`
+  ALTER TABLE "mano"."Report"
+  ADD COLUMN IF NOT EXISTS "collaborations" text[],
+  DROP COLUMN IF EXISTS "collaboration";
+`);