+++ api/src/db/migrations/index.js
@@ -5,4 +5,5 @@ require("./2021-09-10_add-maraude-in-report");
 require("./2021-09-10_team-night-session.js");
 require("./2021-10-11_add-out-of-active-list");
 require("./2021-10-01-custom-fields-obs.js");
+require("./2021-11-05_multiple-collaborations");
 require("./2021-10-22_encrypted-verification-key");